+++ .gitignore
@@ -7,3 +7,4 @@
 /common/__pycache__
 /fno/__pycache__
 /analyser/__pycache__
+/notification/__pycache__
