+++ common/constants.py
@@ -23,6 +23,13 @@ NO_OF_STOCKS = 1
 INTRADAY_SLEEP_TIME = 301
 
 
+#NOTIFICATION CONSTANTS
+
+TELEGRAM_TOKEN = '7042349293:AAGW0-OzOwfvbKdkuM6G40UfcXIHcs_YJwk' 
+TELEGRAM_CHAT_ID = "1462841143"
+TELEGRAM_URL = 'https://api.telegram.org/bot'
+
+
 NseOptionChainURL = "https://www.nseindia.com/option-chain"
 
 # Opstra Data Collection Constants
